+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0 - 2026-04-25
+
+### Features
+
+* The http.webhook alert method can now be configured to send arbitrary
+  headers (e.g. auth tokens).
+
 ## 1.0.0 - 2026-02-26
 
 ### Breaking changes